Replaced some numbers with constants

+++ src/core/data_source.rs
@@ -57,14 +57,14 @@ impl DataSource {
 
     pub fn get_flip_slice_twist_depth_mod_3(&self, flip: u16, twist: u16, slice_sorted: u16) -> u8 {
         let slice = slice_sorted / 24;
-        let flip_slice = ((NFLIP * slice as usize) + (flip as usize)) ;
+        let flip_slice = ((NFLIP * (slice as usize)) + (flip as usize)) ;
 
-        let class_index = self.flip_slice_source.flip_slice_class_index[flip_slice];
+        let class_index = self.flip_slice_source.flip_slice_class_index[flip_slice] as usize;
         let flip_slice_sym = self.flip_slice_source.flip_slice_symmetry[flip_slice];
 
-        let twist_conj = self.moves_source.get_twist_conj(twist, flip_slice_sym);
+        let twist_conj = self.moves_source.get_twist_conj(twist, flip_slice_sym) as usize;
 
-        let ix = (NTWIST * class_index + twist_conj) as usize;
+        let ix = (NTWIST * class_index + twist_conj);
 
         let mut y = self.phase_1_pruning[ix / 16];
         y >>= (ix % 16) * 2;
